backend/pyproject.toml
[tool.poetry]
name = "chessapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask-migrate = "^4.0.0"
flask-sqlalchemy = "^3.0.2"
psycopg2-binary = "^2.9.5"
flask-marshmallow = "^0.14.0"
flask-restful = "^0.3.9"
marshmallow-sqlalchemy = "^0.28.1"
uuid = "^1.30"
sqlalchemy-utils = "^0.38.3"
passlib = "^1.7.4"
argon2-cffi = "^21.3.0"
python-dotenv = "^0.21.0"
flask = "^2.2.2"
pytest-flask = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
